+++ src/components/MessageCard/CreatedAt/CreatedAt.module.css
@@ -1,0 +1,4 @@
+.CreatedAt {
+  font-size: 1.2rem;
+  color: var(--Gray-40);
+}
