+++ src/components/MessageCard/MessageCard.js
@@ -1,20 +1,14 @@
 import { useState } from "react";
 
-import mockData from "../../assets/mock.json";
 import MessageCardContainer from "../MessageCardContainer/MessageCardContainer";
 import MessageCardModal from "../MessageCardModal/MessageCardModal";
-import RelationshipTag from "../RelationshipTag/RelationshipTag";
 
+import CreatedAt from "./CreatedAt/CreatedAt";
 import styles from "./MessageCard.module.css";
 import SenderInfo from "./SenderInfo/SenderInfo";
 
 const MessageCard = ({ message }) => {
   const [messageCardModalOpen, setMessageCardModalOpen] = useState(false);
-
-  const CREATED_AT = new Date(mockData.createdAt);
-  const CREATED_YEAR = CREATED_AT.getFullYear();
-  const CREATED_MONTH = CREATED_AT.getMonth();
-  const CREATED_DATE = CREATED_AT.getDate();
 
   const handleClick = () => {
     setMessageCardModalOpen(true);
@@ -25,9 +19,7 @@ const MessageCard = ({ message }) => {
       <SenderInfo message={message} />
       <div className={styles.DividingLine}></div>
       <div className={styles.CardContent}>{message.content}</div>
-      <p
-        className={styles.CreatedAt}
-      >{`${CREATED_YEAR}.${CREATED_MONTH}.${CREATED_DATE}`}</p>
+      <CreatedAt createdAt={message.createdAt} />
       {messageCardModalOpen && (
         <MessageCardModal
           message={message}
