+++ src/components/MessageCard/MessageCard.module.css
@@ -16,8 +16,3 @@
   height: 10.6rem;
   overflow: hidden;
 }
-
-.CreatedAt {
-  font-size: 1.2rem;
-  color: var(--Gray-40);
-}
