+++ src/components/MessageCardModal/MessageCardModal.js
@@ -1,3 +1,4 @@
+import CreatedAt from "../MessageCard/CreatedAt/CreatedAt";
 import SenderInfo from "../MessageCard/SenderInfo/SenderInfo";
 
 import styles from "./MessageCardModal.module.css";
@@ -6,7 +7,10 @@ const MessageCardModal = ({ message }) => {
   return (
     <div className={styles.ModalBackground}>
       <div className={styles.MessageModalBox}>
-        <SenderInfo message={message} />
+        <div className={styles.MessageModalSender}>
+          <SenderInfo message={message} />
+          <CreatedAt createdAt={message.createdAt} />
+        </div>
       </div>
     </div>
   );
