+++ src/components/MessageCardModal/MessageCardModal.module.css
@@ -13,3 +13,13 @@
   margin: 0 auto;
   padding: 4rem;
 }
+
+.MessageModalSender {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+
+  p {
+    font-size: 1.4rem;
+  }
+}
